+++ skrub/__init__.py
@@ -4,7 +4,7 @@ skrub: Prepping tables for machine learning.
 from pathlib import Path as _Path
 
 from ._check_dependencies import check_dependencies
-from ._datetime_encoder import DatetimeEncoder
+from ._datetime_encoder import DatetimeEncoder, to_datetime
 from ._deduplicate import compute_ngram_distance, deduplicate
 from ._fuzzy_join import fuzzy_join
 from ._gap_encoder import GapEncoder
@@ -32,4 +32,5 @@ __all__ = [
     "TargetEncoder",
     "deduplicate",
     "compute_ngram_distance",
+    "to_datetime",
 ]